+++ cmd/frontend/main.go
@@ -4,10 +4,11 @@ package main
 import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/shared"
 	"github.com/sourcegraph/sourcegraph/internal/sanitycheck"
-	_ "github.com/sourcegraph/sourcegraph/ui/assets/dist"
+	"github.com/sourcegraph/sourcegraph/ui/assets"
 )
 
 func main() {
+	assets.Init()
 	sanitycheck.Pass()
 	shared.FrontendMain(nil)
 }
