+++ cmd/server/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
     deps = [
         "//cmd/server/shared",
         "//internal/sanitycheck",
-        "//ui/assets/dist",
+        "//ui/assets",
     ],
 )
