+++ ui/assets/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "assets",
     srcs = [
         "assets.go",
-        "dev.go",
         "doc.go",
         "manifest.go",
     ],
