+++ ui/assets/assets.go
@@ -1,10 +1,45 @@
 package assets
 
 import (
+	"encoding/json"
+	"fmt"
 	"net/http"
+	"os"
+	"path/filepath"
+	"strings"
 
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
+
+// DefaultAssetPath is the default path where assets should be loaded from. It is primarily used when
+//
+// * AssetPath is empty
+// * env var SRC_ASSETS_DIR is empty
+//
+// This is the default value where assets are created when you run //client/web/dist:copy_bundle
+const DefaultAssetPath = "/assets-dist"
+
+// AssetsPath is absolute path where assets should be loaded from.
+// * During init if it's value is empty the value of the environment variable `SRC_ASSETS_DIR` is used
+// * If the environment variable is ALSO empty, value of the constant `DefaultAssetPath` is used
+var assetsPath = ""
+
+func Init() {
+	// If AssetsPath is empty try:
+	// * Getting the env var value of `SRC_ASSETS_DIR`
+	// * Settle for DefaultAssetPath
+	if assetsPath == "" {
+		path := os.Getenv("SRC_ASSETS_DIR")
+		if path == "" {
+			path = DefaultAssetPath
+		}
+		if !strings.HasPrefix(path, "/") {
+			panic(fmt.Sprintf("SRC_ASSETS_DIR %q is not a absolute path", path))
+		}
+		assetsPath = path
+	}
+	UseAssetsProviderForPath(assetsPath)
+}
 
 // AssetsProvider abstracts accessing assets and the web build manifest.
 // One implementation must be explicitly set in the main.go using
@@ -32,4 +67,52 @@ func (p FailingAssetsProvider) LoadWebBuildManifest() (*WebBuildManifest, error)
 
 func (p FailingAssetsProvider) Assets() http.FileSystem {
 	panic("assets are not configured for this binary, please see ui/assets/doc.go")
+}
+
+// UseDevAssetsProvider installs the development variant of the DirProvider
+// which expects assets to be generated on the fly by an external web builder process.
+func UseDevAssetsProvider() {
+	// When we're using the dev asset provider we expect to be in the monorepo, therefore we use a relative path
+	devAssetsDir := "client/web/dist"
+	Provider = DirProvider{dir: devAssetsDir, assets: http.Dir(devAssetsDir)}
+}
+
+func UseAssetsProviderForPath(path string) {
+	Provider = DirProvider{dir: path, assets: http.Dir(path)}
+}
+
+// DirProvider is the development variant of the UseDevAssetsProvider
+// which expects assets to be generated on the fly by an external web builder process.
+type DirProvider struct {
+	dir    string
+	assets http.FileSystem
+}
+
+func (p DirProvider) LoadWebBuildManifest() (*WebBuildManifest, error) {
+	return loadWebBuildManifest(p.dir)
+}
+
+func (p DirProvider) Assets() http.FileSystem {
+	return p.assets
+}
+
+var MockLoadWebBuildManifest func() (*WebBuildManifest, error)
+
+// loadWebBuildManifest uses a web builder manifest to extract hashed bundle names to
+// serve to the client. In dev mode, we load this file from disk on demand, so it doesn't
+// have to exist at compile time, to avoid a build dependency between frontend
+// and client.
+func loadWebBuildManifest(rootDir string) (m *WebBuildManifest, err error) {
+	if MockLoadWebBuildManifest != nil {
+		return MockLoadWebBuildManifest()
+	}
+
+	manifestContent, err := os.ReadFile(filepath.Join(rootDir, "web.manifest.json"))
+	if err != nil {
+		return nil, errors.Wrap(err, "loading web build manifest file from disk")
+	}
+	if err := json.Unmarshal(manifestContent, &m); err != nil {
+		return nil, errors.Wrap(err, "parsing manifest json")
+	}
+	return m, nil
 }
